Move acceptance test from saved into tests
